Update vector.js
long story short the original version of the rotate function queried the magnitude between updating x and y causing a very weird shape to occur. instead of being circular it sort of looked like a cushion or dog biscuit kinda.

now it's possible to rotate about a point to generate a clean circle.